+++ test/mock-api/sharing/get_asset_share.spec.ts
@@ -1,0 +1,127 @@
+import crypto from 'crypto';
+import { createClient, findWireMockRequest } from '../utils/utils';
+import { expect } from '@jest/globals';
+import { AssetType } from '@smartsheet/sharing';
+import {
+  TEST_SHARE_ID,
+  TEST_ASSET_ID,
+  EXPECTED_SHARE_ALL_PROPERTIES,
+  EXPECTED_SHARE_REQUIRED_PROPERTIES,
+  EXPECTED_SHEET_ASSET_QUERY,
+  ERROR_500_STATUS_CODE,
+  ERROR_500_MESSAGE,
+  ERROR_400_STATUS_CODE,
+  ERROR_400_MESSAGE,
+} from './common_test_constants';
+
+describe('Sharing - getAssetShare endpoint tests', () => {
+  const client = createClient();
+
+  it('getAssetShare generated url is correct', async () => {
+    const requestId = crypto.randomUUID();
+    const options = {
+      shareId: TEST_SHARE_ID,
+      queryParameters: {
+        assetType: AssetType.SHEET,
+        assetId: TEST_ASSET_ID,
+      },
+      customProperties: {
+        'x-request-id': requestId,
+        'x-test-name': '/sharing/get-asset-share/all-response-body-properties',
+      },
+    };
+    await client.sharing.getAssetShare(options);
+    const matchedRequest = await findWireMockRequest(requestId);
+
+    const parsedUrl = new URL(matchedRequest.absoluteUrl);
+    expect(parsedUrl.pathname).toEqual(`/2.0/shares/${TEST_SHARE_ID}`);
+    expect(matchedRequest.method).toEqual('GET');
+
+    const queryParamsObject = Object.fromEntries(parsedUrl.searchParams);
+    expect(queryParamsObject).toEqual(EXPECTED_SHEET_ASSET_QUERY);
+  });
+
+  it('getAssetShare all response body properties', async () => {
+    const requestId = crypto.randomUUID();
+    const options = {
+      shareId: TEST_SHARE_ID,
+      queryParameters: {
+        assetType: AssetType.SHEET,
+        assetId: TEST_ASSET_ID,
+      },
+      customProperties: {
+        'x-request-id': requestId,
+        'x-test-name': '/sharing/get-asset-share/all-response-body-properties',
+      },
+    };
+    const response = await client.sharing.getAssetShare(options);
+    const matchedRequest = await findWireMockRequest(requestId);
+
+    expect(matchedRequest.body).toEqual('');
+    expect(response).toEqual(EXPECTED_SHARE_ALL_PROPERTIES);
+  });
+
+  it('getAssetShare required response body properties', async () => {
+    const requestId = crypto.randomUUID();
+    const options = {
+      shareId: TEST_SHARE_ID,
+      queryParameters: {
+        assetType: AssetType.SHEET,
+        assetId: TEST_ASSET_ID,
+      },
+      customProperties: {
+        'x-request-id': requestId,
+        'x-test-name': '/sharing/get-asset-share/required-response-body-properties',
+      },
+    };
+    const response = await client.sharing.getAssetShare(options);
+    const matchedRequest = await findWireMockRequest(requestId);
+
+    expect(matchedRequest.body).toEqual('');
+    expect(response).toEqual(EXPECTED_SHARE_REQUIRED_PROPERTIES);
+  });
+
+  it('getAssetShare error 400 response', async () => {
+    const requestId = crypto.randomUUID();
+    const options = {
+      shareId: TEST_SHARE_ID,
+      queryParameters: {
+        assetType: AssetType.SHEET,
+        assetId: TEST_ASSET_ID,
+      },
+      customProperties: {
+        'x-request-id': requestId,
+        'x-test-name': '/errors/400-response',
+      },
+    };
+    try {
+      await client.sharing.getAssetShare(options);
+      expect(true).toBe(false); // Expected an error to be thrown
+    } catch (error: any) {
+      expect(error.statusCode).toBe(ERROR_400_STATUS_CODE);
+      expect(error.message).toBe(ERROR_400_MESSAGE);
+    }
+  });
+
+  it('getAssetShare error 500 response', async () => {
+    const requestId = crypto.randomUUID();
+    const options = {
+      shareId: TEST_SHARE_ID,
+      queryParameters: {
+        assetType: AssetType.SHEET,
+        assetId: TEST_ASSET_ID,
+      },
+      customProperties: {
+        'x-request-id': requestId,
+        'x-test-name': '/errors/500-response',
+      },
+    };
+    try {
+      await client.sharing.getAssetShare(options);
+      expect(true).toBe(false); // Expected an error to be thrown
+    } catch (error: any) {
+      expect(error.statusCode).toBe(ERROR_500_STATUS_CODE);
+      expect(error.message).toBe(ERROR_500_MESSAGE);
+    }
+  });
+});